ビルドエラー対応 dbz.business

+++ DB/DBZ/standard/business/src/main/java/jp/co/ndensan/reams/db/dbz/business/core/basic/NinteiChosaScheduleMemoBuilder.java
@@ -6,6 +6,7 @@
 package jp.co.ndensan.reams.db.dbz.business.core.basic;
 
 import static java.util.Objects.requireNonNull;
+import jp.co.ndensan.reams.db.dbz.entity.basic.DbT5222NinteiChosaScheduleMemoEntity;
 import jp.co.ndensan.reams.ur.urz.definition.enumeratedtype.message.UrSystemErrorMessages;
 import jp.co.ndensan.reams.uz.uza.biz.Code;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleDate;
@@ -82,7 +83,7 @@ public class NinteiChosaScheduleMemoBuilder {
      */
     public NinteiChosaScheduleMemoBuilder set連番(Decimal 連番) {
         requireNonNull(連番, UrSystemErrorMessages.値がnull.getReplacedMessage("連番"));
-        entity.setRemban(連番);
+        entity.setRemban(連番.intValue());
         return this;
     }
